Move factories to test modules
