+++ woeip/apps/air_quality/tests/factories.py
@@ -5,7 +5,7 @@ import factory.fuzzy
 from django.contrib.gis import geos
 
 from woeip.apps.air_quality.models import Data, Device, Route, Sensor, Session, SessionData
-from woeip.apps.core.factories import UserFactory
+from woeip.apps.core.tests.factories import UserFactory
 
 
 class FuzzyLineString(factory.fuzzy.BaseFuzzyAttribute):
